M11606: Move PositiveIntegerField from bleaching serializer to fields module
