+++ src/api/ingest/__init__.py
@@ -1,6 +1,7 @@
 from .serializers import *
 from .benthiclit_serializer import *
 from .benthicpit_serializer import *
+from .benthicpqt_serializer import *
 from .bleaching_serializer import *
 from .fishbelt_serializer import *
 from .habitatcomplexity_serializer import *
