+++ src/api/ingest/bleaching_serializer.py
@@ -1,8 +1,8 @@
 from django.utils.translation import gettext_lazy as _
-from rest_framework import fields, serializers
+from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
 
-from ..fields import LazyChoiceField
+from ..fields import LazyChoiceField, PositiveIntegerField
 from ..models import BLEACHINGQC_PROTOCOL, BenthicAttribute, GrowthForm
 from .choices import (
     build_choices,
@@ -14,36 +14,6 @@ from .choices import (
 from .serializers import CollectRecordCSVListSerializer, CollectRecordCSVSerializer
 
 __all__ = ["BleachingCSVSerializer"]
-
-
-class PositiveIntegerField(fields.Field):
-    default_error_messages = {
-        "min_value": _("Ensure this value is greater than or equal to 0.")
-    }
-
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.allow_null = True
-        self.required = False
-        self.allow_blank = True
-        self.default = kwargs.get("default", 0)
-
-    def to_internal_value(self, value):
-        try:
-            val = int(value)
-        except (TypeError, ValueError):
-            val = self.default
-
-        if val is not None and val < 0:
-            self.fail("min_value")
-
-        return val
-
-    def to_representation(self, value):
-        try:
-            return int(value)
-        except (TypeError, ValueError):
-            return 0
 
 
 def growth_form_choices():
